+++ frontend/src/app/components/register/register.component.ts
@@ -1,30 +1,10 @@
-import { Component, Injectable, OnInit } from '@angular/core';
-import { UserService } from '../../services/user/user.service';
+import { Component, OnInit } from '@angular/core';
 import { ReactiveFormsModule, FormControl, FormGroup, Validators, ValidatorFn, AbstractControl, ValidationErrors, AsyncValidator} from '@angular/forms';
 import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
 import { NgIf, NgOptimizedImage } from '@angular/common';
 import { AuthService } from '../../services/auth/auth.service';
-import { Observable, catchError, map, of } from 'rxjs';
-
-export function forbiddenNameValidator(nameRe: RegExp): ValidatorFn {
-	return (control: AbstractControl): ValidationErrors | null => {
-		const forbidden = nameRe.test(control.value);
-		return forbidden ? {forbiddenName: {value: control.value}} : null;
-	};
-}
-
-@Injectable({providedIn: 'root'})
-export class UniqueNameValidator implements AsyncValidator {
-	constructor(private authService: AuthService) {}
-
-	validate(control: AbstractControl): Observable<ValidationErrors | null> {
-		return this.authService.isNameTaken(control.value).pipe(
-			map((isTaken) => (isTaken ? {uniqueName: true} : null)),
-			catchError(() => of(null)),
-		);
-	}
-}
+import { UniqueNameValidator, forbiddenNameValidator } from '../../services/validator/name-validator.service';
 
 @Component({
   selector: 'app-register',
